refactor: 레시피 등록페이지 썸네일 redux 적용 및 styled-components 분리(#208)
